Swaps LS,LS2 for LS1,LS
Global aliases add '| less -S' to the end of commands:

e.g.:

    seq 100 LS

expands (now) to:

    seq 100 |& less -S

Idea of this commit is that usually you want the shorter global alias to
also capture stderr.  So, the old LS2 is now the "default" LS, but the
old LS behavior can be had via LS1.