Add very-WIP old SSH agent limiter
I use enough SSH keys that I often hit the default "you tried too many
keys" limit (when `ssh` just iterates through the ones known to my
agent). So, I was going to write a wrapper to still have all the keys in
my main agent, but this one would delegate to it while only presenting a
subset of the keys when listing.