Add `chrome-desktops` script
Finds a list in the form:

```
ID <tab> desktop <tab> title
```

For each Chromium window with ID and title with what desktop it's on.
E.g.:

```
0x60004d	7	"Bizarre Spinosaurus makes history as first known swimming dinosaur - Chromium"
0x60004a	7	"Kosmi - Chromium"
0x600047	7	"X keyboard extension - ArchWiki - Chromium"
```

Wrote this when I was causing some crashes with XMonad, so constantly
had to reorganize my Chromium windows (annoying when you have, e.g., 68
open as I currently do).