Rework "dotted directory" changing in Vim
Rsync has a great `-R` which lets you put a superfluous `/./` in a
source path to make other filenames relative to that point in the path
hierarchy.

I'm not sure whether something changed in Vim, or if I never used it,
but my function for doing a similar thing in Vim was broken: the
filename reported by `expand("<afile>")` no longer contained the `/./`.
Instead, if present, use the arglist version of the current window's
filename.