First pass at `with-known-host` script
Should be more versatile than `ssh-ngrok` script.  Instead of creating a
custom `known_hosts` file, it overmounts `~/.ssh` using the mnt
namespace and adds the desired host to `~/.ssh/known_hosts` directly.
But, there are currently a couple issues with the fact that  current
shell context is lost when it spawns the additional layers (`unshare`
and `nsenter`).