Add `$ONLY` and `$SKIP` vars for `git.encryptedjs`
Both are whitespace- and/or comma-separated "lists".

Putting a stage in `$SKIP` removes it from consideration.

If `$ONLY` is set, it implicitly skips all other stages.  If a stage is
(pointlessly) listed in both, `$SKIP` takes precedence.

If `"puts"` is not included in `$ONLY`, but any of the default stages
is, it will be added back. (Otherwise, there won't be any output.)