Beautiful mess of `plug-keyboard` scripts
Can't just set the X11 properties of the keyboard when it's plugged in,
because X hasn't claimed the device yet.

So, tried to double-fork in various ways and spawning off other
processes to wait around until X claims it.

But, turns out the child process is getting killed because everything's
in a Systemd `cgroup`.

So, gonna start from scratch (basically).