Limit `<Alt-w>` error message when file is changed
The standard error message reports the location in VimL where the error
occurred, but that's overkill for the "hey, this file is modified"
failure of trying to close a window. So, suppress normal error
reporting, but replace it with manual, `:uns[ilent]`
`ErrorMsg`-highlighted echo. (Most importantly, this avoids the "Hit
Enter to continue" prompt.)