Add `git devbranch` for `dev`-related aliases
Previously, these were all hardcoded to `dev` (the name of the branch
we used where I got in the habit of using `dev` for the basis of lots
of actions). At another $job, though, we used `development`. Instead
of creating all-new versions (`dd = diff development`, etc.), allow
the name of the branch to be conigurable (via `git config dev.branch
development`).