Add git `last-touched` + `auto-fixup` (`af`)
`git last-touched` finds when the changed lines (either working dir
changes or staged) were last touched, either as a list of changes (kind
of like `git blame --line-porcelain`) or a list of commits.

`git auto-fixup` (alias `af`, for `gaf`) presents the output of `git
last-touched` in a way that makes it easy to commit as a `--fixup`
commit.