Add `-i` = `sed`-style in-place editing wrapper
Allows `perl`/`sed` style `-i`: run the command with file on stdin, but use
stdout as the replacement contents. Uses `.-i` as the default backup filename.
Doesn't replace the file if the command fails.

Examples:

```sh
$ echo aaa > filename
$ -i filename sed -e s/a/b/ -e s/a/l/
$ cat filename
bla
$ ls filename*
filename
```

Was annoyed that `sed -i -e cmd1 -e cmd2 filename` is not portable.

On BSD-descended `sed` (e.g., macOS built-in), `-i` is not interpreted as
`-i` + empty extension. Instead, the next argument is always used.

For GNU `sed`, `-i` is interpreted as `-i` + empty extension, but it's not
allowed to split the extension to the next argument (it must be "glommed").